src/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
distlib==0.3.1
Django==3.0.8
Django-Admin-Object-Actions==0.1.7
django-crum==0.7.6
django-environ==0.4.5
django-filter==2.3.0
django-graphql-jwt==0.3.0
filelock==3.0.12
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
mysqlclient==2.0.1
pipenv==2020.6.2
promise==2.3
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.28
virtualenv-clone==0.5.4
